add arkiv link, move utils
