--- src/llmcompressor/modifiers/quantization/gptq/utils/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-
-from .gptq_quantize import *
